backend/requirements.txt
# AyuAahar Backend Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
